refactor(routine): decompose RoutineSettingsSection into sub-components
RoutineSettingsSection.jsx went from 961 LOC to 203 LOC by extracting
8 presentational sub-components into src/modules/routine/components/settings/.
Pure structural refactor — behavior and markup are identical to main.

New files (settings/):
- HabitForm.jsx (189 LOC) — new/edit habit form composition.
- ReminderPresets.jsx (106 LOC) — reminder preset chips + per-time rows.
- WeekdayPicker.jsx (45 LOC)  — weekly-recurrence weekday picker,
  wrapped in React.memo (pure props-driven).
- TagsSection.jsx (111 LOC) — tag CRUD (create/rename/delete) with
  inline-edit state localized to the section.
- CategoriesSection.jsx (177 LOC) — category CRUD with its own
  ConfirmDialog for deletes.
- HabitListItem.jsx (116 LOC) — one draggable habit row (↑↓, Деталі,
  Змінити, В архів, Видалити). React.memo so editing one habit no
  longer re-renders every row.
- ActiveHabitsSection.jsx (136 LOC) — search + ordered list of active
  habits, local dragId/query state.
- ArchivedHabitsSection.jsx (58 LOC) — archive list with restore/
  delete actions.

RoutineSettingsSection.jsx itself now only owns cross-cutting state
that is shared between sections: editingId (used by HabitForm and
ActiveHabitsSection), detailHabitId (for HabitDetailSheet), and the
two globally-triggered ConfirmDialogs (habit delete + import confirm).

Local state that is scoped to a single section was moved into that
section:
- habitListQuery, dragId        → ActiveHabitsSection
- editingTagId, tagSavedRef…    → TagsSection
- editingCatId, deleteCatPending → CategoriesSection

Tests 543/543 pass, lint + typecheck clean. No behavior change.
